Update BuildAndPack.yml to simplify workflow
The changes made in this commit include the removal of auto-generated comments, removal of 'master' branch in the 'push' trigger, adjustment of the dotnet version setup, and the removal of steps related to globaljson, caching, and the build command. These adjustments have been made to streamline the workflow process, making it less complex and easier to manage.